Commute_Score/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
jinja2>=3.1.3
pydantic>=2.7.1
httpx>=0.27.0
python-dotenv>=1.0.1